requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Django==4.2.5
djangorestframework==3.14.0
flake8-isort==6.0.0
idna==3.4
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
pycodestyle==2.11.0
pyflakes==3.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
